Comments - Defect Number: 3374; Defect RecordID: 3370; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: E00092 - Right to Present; Defect Description: Hi Nick,
Please find the following questions and let us know your response.
1)Person Detail:
    a)Compensation tab:
i)As per spec document- "A rate is required" for resources with "RighttoPresent" status. We are assuming that user can select any Compensation type to assign a rate to resource. Please Confirm.
confirmed
ii)'Practice Area' is required to add a Compensation record to resource. Please let us know if we need to follow the same rule for "RighttoPresent" status also.
Ellen- Practice Area should be required forRighttoPresentresources, as it is used as a pivot for other behavior and would greatly simplify the number of exceptions wed need to add to the existing logic (read on).
iii)Currently, when user selects 1099/Hourly or 1099/POR Compensation type, Vendor Indicator filed will appear and it is a mandatory field. Please let us know if we need to follow the same rule for "RighttoPresent" status also.
confirmed
    b)Recruiting Metrics:
i)According to document, we need to add Third Party to Recruiter drop down. We are assuming that, Third Party value is available for all statuses.
confirmed
ii)When user selects Third Party from Recruiter drop down, then we need to show only Vendors (we need to exclude all the other sources from drop down) in Source drop down.
Yes, lets do that. The source list would mirror the vendor list and would bepre-populatedwith the Vendor selected on the Compensation tab if the Compensation Type was 1099.otherwise the user needs to select from the list.
    c) We are assuming that user can able to move Resource with RighttoPresent status to any status (Active, Contingent, Terminated ). Please confirm.
Correct. This would be the standard workflow that once the SOW had been signed for a project, the Person would be converted to Active status, which would then trigger all of the standard edits for Active Persons.
    d) Let us consider a Resource with RighttoPresent status and Hire Date as 08/30/2017. When user move this to Active, then we need to populate 08/30/2017 as Hire Date and user can able to change the date if needed. Correct?
confirmed
    e) Do we need to include RighttoPresent status in Configuration > Company Resources > Persons > Filters > Person Status.
confirmed. Please add this value.
 Human Capital Report
    a)New Hire Report
i) As per spec we need to show resources with RighttoPresent status in this report. However, report is generated according to the selected filters and Resources with Righttopresent status may not be available in the report asPractice Area andHire Date are not required for Righttopresent Resources.
Ellen I think we change the requirement on Hire Date to be required (it is currently NOT required per spec). Please confirm. Either that, or we need to modify the behavior of this report.but there is no clean alternate solution in my mind other than showing ALLRighttoPresentpersons.

Termination Report
i) As per spec we need to show resources with RighttoPresent status in this report. However, report is generated according to the selected filters and Resources with Righttopresent status may not be available in the report as Practice Area is not requiredfor Righttopresent Resources and Termination Date and Terminations reason is not available.

@Ellen a Termination Date assumes (rightly so I think) a Hire Date..currently the Termination Date field is not available until a Hire Date is entered. In the case of aRighttoPresent, these dates are probably better referred to as RighttoPresentStart Date and RighttoPresentEnd Date.but I am not advocating changing the titles of the fields based on the new Status type (seems messy). Id say that we requi